Wordpress installation recipe.
This will install WP 3.0 into /var/lib/wordpress/XXXX where XXXX is defined
via databag for the specific host. The installation will also configure MySQL
with a new database for the specific installation and also create a nginx
configuration file in /etc/nginx/servers.

What is installed and where it is installed depends on a databag that should
be installed in the chef master repo. The databag should be called 'wordpress.json'
and have the form, e.g.:

[
 { "id": "wordpress_on_some_machine",
   "host": "host1.example.com",
   "wp": [
     [ "wp_one", "wp_one.example.com" ],
     [ "wp_two", "wp_two.example.com" ],
  ]
 },
 { "id": "wordpress_on_some_other_machine",
   "host": "host2.example.com",
   "wp": [
     [ "wp_three", "wp_three.example.com" ],
     [ "wp_four", "wp_four.example.com" ],
  ]
 }
]

Where:
 id
  is required but not used. Don't use any '.' or symbols.
 host
  is also required and is the fqdn of the node.
 wp is an array of arrays with name, hostname(, action) for the WP
  installation. Name defines the mysql user and database name.
  Name also is the directory in /var/lib/wordpress where the install
  is placed. The hostname is the nginx server name under which the
  installation should be reachable. SSL + HTTP ports are config'ed.
  Action is optional and defaults to 'create' but can be 'delete' to
  remove the installation. WARNING: the database is also trashed.
  IF ACTION == DELETE THEN DELETE ALL MySQL DATA AND DIRECTORIES